lambdas/authentication/requirements.txt
bcrypt==4.0.1
boto3==1.29.3
botocore==1.32.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
flake8==6.1.0
idna==3.6
jmespath==1.0.1
mccabe==0.7.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
python-dateutil==2.8.2
requests==2.31.0
s3transfer==0.7.0
sentry-sdk==1.39.2
six==1.16.0
urllib3==1.26.18
